i actually dont know if this is correct.
i actually dont know if this is correct. As everyone already knows i had an issue with tut3 so obviously the challenge was challenging. The benchmark was being killed before even starting probably because i was stingy with the vm resources (i gave it 2 cores and 1500mb RAM) So i reduced the Ns and it worked for a bit before being killed again so idk atp *insert sad face*.